+++ rxrepo-core/src/main/java/com/slimgears/rxrepo/query/provider/HasLimit.java
@@ -3,5 +3,5 @@ package com.slimgears.rxrepo.query.provider;
 import javax.annotation.Nullable;
 
 public interface HasLimit {
-    @Nullable Integer limit();
+    @Nullable Long limit();
 }
