+++ rxrepo-core/src/main/java/com/slimgears/rxrepo/query/provider/HasPagination.java
@@ -3,5 +3,5 @@ package com.slimgears.rxrepo.query.provider;
 import javax.annotation.Nullable;
 
 public interface HasPagination extends HasLimit {
-    @Nullable Integer skip();
+    @Nullable Long skip();
 }
